Labfiles/03-language/Python/clock-client/requirements.txt
python-dotenv
azure-ai-language-conversations
